update: rename packages in commands package
