+++ src/main/java/tech/araopj/springpitzzahhbot/DevBot.java
@@ -1,20 +1,20 @@
 package tech.araopj.springpitzzahhbot;
 
 import tech.araopj.springpitzzahhbot.services.slash_commands.ConfessionService;
-import tech.araopj.springpitzzahhbot.commands.slash_command.commands.joke.ApproveJoke;
-import tech.araopj.springpitzzahhbot.commands.slash_command.commands.joke.SubmitJoke;
+import tech.araopj.springpitzzahhbot.commands.slash_commands.commands.joke.ApproveJoke;
+import tech.araopj.springpitzzahhbot.commands.slash_commands.commands.joke.SubmitJoke;
 import tech.araopj.springpitzzahhbot.services.slash_commands.JokesService;
 import tech.araopj.springpitzzahhbot.services.slash_commands.GameService;
-import tech.araopj.springpitzzahhbot.commands.slash_command.commands.Confession;
-import tech.araopj.springpitzzahhbot.commands.slash_command.commands.joke.GetJoke;
-import tech.araopj.springpitzzahhbot.commands.chat_command.commands.FormatChatCommand;
+import tech.araopj.springpitzzahhbot.commands.slash_commands.commands.Confession;
+import tech.araopj.springpitzzahhbot.commands.slash_commands.commands.joke.GetJoke;
+import tech.araopj.springpitzzahhbot.commands.chat_commands.commands.FormatChatCommand;
 import tech.araopj.springpitzzahhbot.services.configs.MessageCheckerService;
-import tech.araopj.springpitzzahhbot.commands.chat_command.commands.HelpChatCommand;
-import tech.araopj.springpitzzahhbot.commands.chat_command.commands.PingChatCommand;
-import tech.araopj.springpitzzahhbot.commands.slash_command.SlashCommandManager;
+import tech.araopj.springpitzzahhbot.commands.chat_commands.commands.HelpChatCommand;
+import tech.araopj.springpitzzahhbot.commands.chat_commands.commands.PingChatCommand;
+import tech.araopj.springpitzzahhbot.commands.slash_commands.SlashCommandManager;
 import tech.araopj.springpitzzahhbot.services.configs.ViolationService;
-import tech.araopj.springpitzzahhbot.commands.slash_command.commands.Play;
-import tech.araopj.springpitzzahhbot.commands.chat_command.ChatCommandManager;
+import tech.araopj.springpitzzahhbot.commands.slash_commands.commands.Play;
+import tech.araopj.springpitzzahhbot.commands.chat_commands.ChatCommandManager;
 import tech.araopj.springpitzzahhbot.services.configs.CategoryService;
 import tech.araopj.springpitzzahhbot.services.configs.ChannelService;
 import tech.araopj.springpitzzahhbot.services.MessageUtilService;
