+++ src/main/java/tech/araopj/springpitzzahhbot/commands/chat_commands/commands/HelpChatCommand.java
@@ -21,12 +21,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package tech.araopj.springpitzzahhbot.commands.chat_command.commands;
+package tech.araopj.springpitzzahhbot.commands.chat_commands.commands;
 
-import tech.araopj.springpitzzahhbot.commands.chat_command.ChatCommandManager;
+import tech.araopj.springpitzzahhbot.commands.chat_commands.ChatCommandManager;
 import tech.araopj.springpitzzahhbot.services.MessageUtilService;
-import tech.araopj.springpitzzahhbot.commands.chat_command.CommandContext;
-import tech.araopj.springpitzzahhbot.commands.chat_command.ChatCommand;
+import tech.araopj.springpitzzahhbot.commands.chat_commands.CommandContext;
+import tech.araopj.springpitzzahhbot.commands.chat_commands.ChatCommand;
 import tech.araopj.springpitzzahhbot.services.CommandsService;
 import org.springframework.stereotype.Component;
