+++ src/main/java/tech/araopj/springpitzzahhbot/commands/slash_commands/commands/joke/ApproveJoke.java
@@ -22,12 +22,12 @@
  * SOFTWARE.
  */
 
-package tech.araopj.springpitzzahhbot.commands.slash_command.commands.joke;
+package tech.araopj.springpitzzahhbot.commands.slash_commands.commands.joke;
 
 import tech.araopj.springpitzzahhbot.services.slash_commands.JokesService;
-import tech.araopj.springpitzzahhbot.commands.slash_command.CommandContext;
+import tech.araopj.springpitzzahhbot.commands.slash_commands.CommandContext;
 import tech.araopj.springpitzzahhbot.services.MessageUtilService;
-import tech.araopj.springpitzzahhbot.commands.slash_command.SlashCommand;
+import tech.araopj.springpitzzahhbot.commands.slash_commands.SlashCommand;
 import net.dv8tion.jda.api.interactions.commands.build.CommandData;
 import net.dv8tion.jda.api.interactions.commands.build.OptionData;
 import net.dv8tion.jda.api.interactions.commands.OptionMapping;
