+++ src/main/java/tech/araopj/springpitzzahhbot/listeners/MessageListener.java
@@ -36,9 +36,9 @@ import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.stereotype.Component;
-import tech.araopj.springpitzzahhbot.commands.chat_command.ChatCommandManager;
+import tech.araopj.springpitzzahhbot.commands.chat_commands.ChatCommandManager;
 import tech.araopj.springpitzzahhbot.services.CommandsService;
-import tech.araopj.springpitzzahhbot.commands.slash_command.commands.Confession;
+import tech.araopj.springpitzzahhbot.commands.slash_commands.commands.Confession;
 import tech.araopj.springpitzzahhbot.services.slash_commands.ConfessionService;
 import tech.araopj.springpitzzahhbot.services.configs.CategoryService;
 import tech.araopj.springpitzzahhbot.services.configs.ChannelService;
